+++ tests/django41.rs
@@ -1,0 +1,55 @@
+//! This is an almost line-by-line translation from the hashers' test from Django 4.1:
+//! https://github.com/django/django/blob/master/tests/auth_tests/test_hashers.py
+//! ...but only for the tests where the iterations differ from Django 1.9.
+
+use djangohashers::*;
+
+#[test]
+#[cfg(feature = "with_pbkdf2")]
+fn test_pbkdf2() {
+    let django = Django {
+        version: DjangoVersion::V4_1,
+    };
+    let encoded = django.make_password_with_settings("lètmein", "seasalt", Algorithm::PBKDF2);
+    assert_eq!(
+        encoded,
+        "pbkdf2_sha256$390000$seasalt$8xBlGd3jVgvJ+92hWPxi5ww0uuAuAnKgC45eudxro7c="
+    );
+    assert!(is_password_usable(&encoded));
+    assert_eq!(check_password("lètmein", &encoded), Ok(true));
+    assert_eq!(check_password("lètmeinz", &encoded), Ok(false));
+    // Blank passwords
+    let blank_encoded = django.make_password_with_settings("", "seasalt", Algorithm::PBKDF2);
+    assert!(blank_encoded.starts_with("pbkdf2_sha256$"));
+    assert!(is_password_usable(&blank_encoded));
+    assert_eq!(check_password("", &blank_encoded), Ok(true));
+    assert_eq!(check_password(" ", &blank_encoded), Ok(false));
+}
+
+#[test]
+#[cfg(feature = "with_pbkdf2")]
+fn test_low_level_pbkdf2() {
+    let django = Django {
+        version: DjangoVersion::V4_1,
+    };
+    let encoded = django.make_password_with_settings("lètmein", "seasalt2", Algorithm::PBKDF2);
+    assert_eq!(
+        encoded,
+        "pbkdf2_sha256$390000$seasalt2$geC/uZ92nRXDSjSxeoiBqYyRcrLzMm8xK3ro1QS1uo8="
+    );
+    assert_eq!(check_password("lètmein", &encoded), Ok(true));
+}
+
+#[test]
+#[cfg(feature = "with_pbkdf2")]
+fn test_low_level_pbkdf2_sha1() {
+    let django = Django {
+        version: DjangoVersion::V4_1,
+    };
+    let encoded = django.make_password_with_settings("lètmein", "seasalt2", Algorithm::PBKDF2SHA1);
+    assert_eq!(
+        encoded,
+        "pbkdf2_sha1$390000$seasalt2$aDapRanzW8aHTz97v2TcfHzWD+I="
+    );
+    assert_eq!(check_password("lètmein", &encoded), Ok(true));
+}